remove utils/common and move content to consolidate/api
